Implement set id logic

+++ app/src/main/java/com/ict/mito/gootravel/spot/select/radar/ui/dialog/SelectSpotBottomSheetFragment.kt
@@ -30,7 +30,7 @@ class SelectSpotBottomSheetFragment : BottomSheetDialogFragment() {
         )
 
         val id: Int = arguments?.getInt("spotId") ?: 0
-        viewmodel.setId(id)
+        viewmodel.setId(id.toLong())
 
         binding?.viewmodel = viewmodel
